pr comments and pydantic updates
